src/backend/requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.14.3
chardet==3.0.4
cryptography==3.3.2
Django==2.2.24
djangorestframework==3.12.4
django-rest-swagger==2.2.0
idna==2.6
ipaddress==1.0.22
psycopg2-binary==2.8.6
pycparser==2.18
pyOpenSSL==17.5.0
pytz==2018.4
requests==2.25.1
six==1.11.0
urllib3==1.26.5
whitenoise==4.1
